uploadVideo: index.js -> upload-video.js
